requirements.txt
absl-py==0.9.0
appnope==0.1.3
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
blinker==1.4
cachetools==3.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
colorama==0.4.3
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
docopt==0.6.2
dpcpp-cpp-rt==2022.1.0
entrypoints==0.3
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.27.2
h5py==2.10.0
idna==2.9
importlib-metadata==1.7.0
intel-cmplr-lic-rt==2022.1.0
intel-openmp==2022.1.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
joblib==0.15.1
jsonpickle==1.4.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.1.1
matplotlib==3.1.3
MiniSom==2.3.0
mkl
mkl-fft
mkl-random
munch==2.5.0
nbformat==5.0.7
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.4.0.42
opt-einsum==3.1.0
packaging==20.4
pandas==1.0.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.8.2
ply==3.11
prompt-toolkit==3.0.4
protobuf
ptyprocess==0.6.0
py-cpuinfo==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==18.1.1
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
sacred==0.8.1
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.1
sip==6.6.1
six==1.14.0
smmap==3.0.4
tbb==2021.6.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.2.0
tensorflow-estimator==2.2.0
tensorflow-probability==0.10.0
termcolor==1.1.0
toml==0.10.2
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
